Middle of work

+++ PN2S/core/models/ModelStatistic.h
@@ -6,10 +6,10 @@ namespace models {
 class ModelStatistic {
 public:
 	double dt;
-	int nCompts;
-	int nModels;
+	size_t nCompts;
+	size_t nModels;
 	ModelStatistic(): nCompts(0), nModels(0), dt(1){}
-	ModelStatistic(double _dt, int _nModel, int _nCompt): dt(_dt), nModels(_nModel), nCompts(_nCompt){}
+	ModelStatistic(double _dt, size_t _nModel, size_t _nCompt): dt(_dt), nModels(_nModel), nCompts(_nCompt){}
 	virtual ~ModelStatistic(){}
 };
